refactor file name, lint issues
