+++ src/helperFunctions/beautifyTaskStatus.ts
@@ -1,4 +1,4 @@
-import task from '../types/task.type';
+import task from '../interfaces/task.type';
 
 const beautifyTaskStatus = (tasks: Array<task>) => {
   let taskList: Array<task> = [];
