+++ src/interfaces/award.type.ts
