+++ src/interfaces/challenge.type.ts
