+++ src/interfaces/level.type.ts
