+++ src/interfaces/user.type.ts
