+++ src/pages/challenges.tsx
@@ -5,8 +5,8 @@ import Active from '@/components/challenges/active';
 import Complete from '@/components/challenges/complete';
 import Accordion from '@/components/Accordion';
 import useFetch from '@/hooks/useFetch';
-import challenge from '@/types/challenge.type';
-import userType from '@/types/user.type';
+import challenge from '@/interfaces/challenge.type';
+import userType from '@/interfaces/user.type';
 import classNames from '@/styles/tasks.module.scss';
 import { CHALLENGES_URL, LOGIN_URL } from '@/components/constants/url';
 import userData from '@/helperFunctions/getUser';
